// server/src/models/types.ts
export type ProductState = 'Nuevo' | 'Oferta' | 'Más vendido' | 'Selección CasaViva' | 'Agotado';
// ...
export interface ProductVariant {
  id: string;
  producto_id: string;
  sku: string;
  nombre_variante: string;
  color_nombre: string;
  color_hex: string;
  stock: number;
  precio_adicional: number;
  imagen_variante?: string;
}

export interface Product {
  id: string;
  slug: string;
  nombre: string;
  descripcion: string;
  descripcion_corta: string;
  categoria_id: string;
  categoria_nombre?: string;
  marca: string;
  precio_base: number;
  precio_anterior?: number;
  descuento_porcentaje?: number;
  imagen_principal: string;
  galeria_imagenes: string[];
  dimensiones: string;
  cuidados: string;
  caracteristicas: string[];
  estado: ProductState;
  activo: boolean;
  calificacion_promedio: number;
  total_resenas: number;
  destacado?: boolean;
  novedad_bajo_100?: boolean;
  variantes: ProductVariant[];
}